Extract ImageJFunctions.wrap to RAIToImagePlus
